+++ CHANGELOG.md
@@ -4,13 +4,19 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 Check [LANGUAGE.md](./LANGUAGE.md) for latest documentation.
 
-## Unreleased
+## 3.0.0 (Unreleased)
+
+### Breaking changes
+
+- Return End of Dialogue object instead of undefined. `{ type: 'end' }`.
+- Variation `shuffle` will do real randomization with no guarantee all items will be visited.
 
 ### Added
 
 - Assignment initializer operator `?=`. It only assigns if variable was not set before.
 - External variables using `@` prefix. They are not included in the save object, so they are not persisted between runs.
 They are useful when dealing with data that belongs to your game and shouldn't be persisted with the dialogue.
+
 
 ### Changed